fix: lazy safetensor patch
